refactor: WebUILottoApplication의 구현 코드들 Service로 추출
