+++ src/main/java/lotto/service/MainService.java
@@ -1,0 +1,4 @@
+package lotto.service;
+
+public class MainService {
+}
