+++ src/main/resources/templates/showLottos.html
