+++ src/main/resources/templates/start.html
@@ -8,7 +8,7 @@
 
     <h1>{{round}} 회 로또</h1>
 
-    <form method="POST" id="lotto" action="/buyLotto">
+    <form method="POST" id="lotto" action="/lottos">
         구입금액을 입력해 주세요.
         <p><input type="number" id="money" name="money">원</p>
